+++ package.json
@@ -1,7 +1,8 @@
 {
     "name": "@asnc/ts_hook",
+    "type": "module",
     "main": "./out/complier/index.js",
-    "version": "0.0.5",
+    "version": "0.0.7",
     "publishConfig": {
         "access": "public",
         "registry": "https://registry.npmjs.org/"
@@ -32,11 +33,13 @@
     "devDependencies": {
         "@swc/core": "^1.3.52",
         "@types/node": "^18.15.3",
-        "ttypescript": "^1.5.15",
         "vite": "^4.1.4",
         "vitest": "^0.29.2"
     },
     "dependencies": {
         "source-map-support": "^0.5.21"
+    },
+    "engines": {
+        "node": ">=16"
     }
 }
