+++ tsconfig.build.json
@@ -1,7 +1,8 @@
 {
     "extends": "./tsconfig.json",
     "compilerOptions": {
+        "target": "ES2022",
         "skipLibCheck": true
     },
-    "exclude": ["**/test","**/__test__", "src/**/*.test.mts", "src/**/*.test.cts", "src/**/*.test.ts"]
+    "exclude": ["**/test", "**/__test__", "src/**/*.test.mts", "src/**/*.test.cts", "src/**/*.test.ts"]
 }
